requirement.txt
numpy==1.21.6
scikit-image==0.18.3
scikit-learn==1.0.1
matplotlib==3.4.3
pandas==1.3.5
pytorch==1.7.1
torchvision==0.8.2
